Handle automation integration test endpoint URLs
The integration tests need to know the endpoint URLs they are testing
against so that they can actually communicate with them and run tests.
We could use an environment variable for this, but the endpoint URL
isn't known until the function is deployed and setting environment
variables across Docker Compose containers is messy.

Instead, this commit uses a shared volume to find the endpoint URL. This
is probably not a _great_ way of doing it (maybe a sqlite database
would've been better?), but its a quick and dirty solution.