Fix response format
AWS API Gateway expects responses sent to it to look like this:

```ruby
{
  :statusCode => $A_STATUS_CODE,
  :body => $JSON_ENCODED_BODY
}.to_json
```

Otherwise, CloudFront sends back `HTTP/2 502 "Internal Server Error"`
(which makes VERY LITTLE sense since (a) this is technically a `HTTP
500` condition, (b) `HTTP 502` is actually "Bad Gateway", not "Internal
Server Error" and (c) CloudFront was actually able to hit the origin, so
the gateway was actually not bad).

I hope this commit helps someone avoid spending 4-5 hours of their live
parsing logs that look good, X-Ray traces that say little and
documentation that should be better.