Fix stubbed responses
Documentation on this is really poor (or I didn't read it well enough).

You don't need to construct the structs yourself (I thought you did
since the AWS SDK calls expect structs of some kind). You can send it a
hash of properties in the expected response that you want stubbed and
AWS will take care of the rest.