Automatically register integration test callbacks
Slack requires us to register our callback URLs with the app before we
use them for OAuth authentication. However, since we are using the
randomly-generated API Gateway endpoint URLs that AWS provides, doing
this for brand-new environments is laborious and will make CI
impossible.

We could use a custom domain for our API Gateway, but this would require
a CloudFront distribution in front of it, which can take up to 45
minutes to provision. We could provision the distribution and attempt to
preserve it whenever we refresh our Lambda functions and their Gateways,
but this comes with two problems:

1. Updating the distribution still takes a while, and
2. This prevents us from having completely pure integration tests, which
can cause DNS-based jankiness and flaky tests later on.

Therefore, I figured that automating the callback URL registration with
Selenium was the easiest way around this. I suspect that Slack will
probably prevent me from doing this at some point, and I'll have to be
careful about respecting rate limits, but it works for now and was
surprisingly easy to do.